chore: #91 move txn into dir
create general alias for sha512- propagate usage;
