+++ src/ledger/txn/mod.rs
@@ -1,12 +1,17 @@
+pub mod utils;
 // imports
 use chrono::prelude::*;
-use ed25519_dalek::{Digest, Sha512};
+use ed25519_dalek::Digest;
 use serde::Serialize;
 use serde_big_array::BigArray;
 use std::fmt;
 // local
 use crate::{
-    ledger::{general::PbKey, txn_pool::TxnMapKey, wallet::Wallet},
+    ledger::{
+        general::{PbKey, Sha512},
+        txn_pool::TxnMapKey,
+        wallet::Wallet,
+    },
     utils::signature::{TxnSignature, TXN_SIGNATURE_CONTEXT},
 };
 pub const TXN_MSG_CTX: &[u8; 6] = b"txn-v0";
@@ -107,8 +112,7 @@ pub struct Txn {
 
 impl Txn {
     /// ## Transaction constructor fxn
-    ///
-    /// Creates a transaction `object`
+    /// Creates a transaction `object`.
     pub fn new(
         pbkey_send: PbKey,
         pbkey_recv: PbKey,
@@ -136,9 +140,9 @@ impl Txn {
         txn
     }
 
-    /// Create and return a new signed transaction.
-    ///
+    /// ## Create and return a new signed transaction.
     /// Receives `Wallet` instance for signing.
+    ///
     /// Uses `Txn::new()` assoc fxn. to construct the txn, and signs the txn with given wallet.
     pub fn new_signed(
         wallet: &Wallet,
@@ -216,41 +220,6 @@ impl Txn {
 
     /////////////////////////////////////////////////////////////////////
     /////////////////////////////// UTILS ///////////////////////////////
-
-    /// Convert transaction struct to bytes - NOT id/hash/message/digest
-    /// TODO: replace `Vec<u8>` - don't allocate
-    pub fn to_bytes(&self) -> Vec<u8> {
-        // serialize to a byte vector
-        serde_json::to_vec(&self).expect("Error serializing txn")
-    }
-
-    /// ## Compute the id (hash digest) of the transaction.
-    /// Converts semantic data for the txn - all non-calculated fields (i.e. excludes `id` and `signature`) into bytes.
-    ///
-    /// Hashes this info and produces a digest - the ID.
-    pub fn calc_id(&self) -> TxnId {
-        let prehash = self.calc_id_sha512_prehash();
-
-        // return the hash digest - the block's id
-        let digest: TxnDigest = prehash.finalize().into();
-        TxnId(digest)
-    }
-    /// ## Calculate the pre-hash struct for the id
-    pub fn calc_id_sha512_prehash(&self) -> ed25519_dalek::Sha512 {
-        // Create a hash digest object which we'll feed the message into:
-        let mut prehash: Sha512 = Sha512::new();
-        // add the txn version
-        prehash.update(TXN_MSG_CTX);
-        // add the txn bytes
-        prehash.update(self.to_bytes());
-        // return the hasher/prehash struct
-        prehash
-    }
-
-    /// ## Create and return a message signature based on the contents of the transaction
-    pub fn calc_signature(&self, wallet: &Wallet) -> TxnSignature {
-        wallet.sign_txn(self)
-    }
 
     /////////////////////////////// UTILS ///////////////////////////////
     /////////////////////////////////////////////////////////////////////
